backend/requirements.txt
fastapi
uvicorn[standard]
gunicorn
python-multipart
pdfminer.six
scikit-learn
transformers
torch
pytest
pydantic
